requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
bokeh==2.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cycler==0.10.0
dask==2.19.0
dask-labextension==2.0.2
decorator==4.4.2
defusedxml==0.6.0
distributed==2.19.0
entrypoints==0.3
HeapDict==1.0.1
idna==2.10
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1607956439537/work
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
jupyterlab-git==0.20.0
kiwisolver @ file:///D:/bld/kiwisolver_1610099948875/work
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy @ file:///D:/bld/numpy_1610324782150/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging==20.4
pandas @ file:///D:/bld/pandas_1610610336238/work
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1610407616132/work
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1608904108784/work
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
scikit-learn @ file:///D:/bld/scikit-learn_1611079981849/work
scipy @ file:///C:/bld/scipy_1609458016927/work
seaborn==0.10.1
Send2Trash==1.5.0
simpervisor==0.3
six==1.14.0
sortedcontainers==2.2.2
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typing-extensions==3.7.4.2
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.4.2
zict==2.0.0
